feat: Save entered title in NewExpense widget
This commit adds functionality to save the entered title in the NewExpense widget. The `_saveTitleInput` method is called when the title text field is changed, and it updates the `_enteredTitle` variable with the new value. This enhancement improves the data handling in the NewExpense widget, allowing the entered title to be used for further processing, such as saving the expense.