feat: Update NewExpense widget to use isScrollControlled for showModalBottomSheet
This commit updates the NewExpense widget to use the `isScrollControlled` parameter for the `showModalBottomSheet` function. By setting `isScrollControlled` to `true`, the bottom sheet will take up the full height of the screen and allow scrolling when necessary. This improvement enhances the user experience by ensuring that all content in the bottom sheet is accessible, even on smaller screens.