feat: Add remove expense functionality to Expenses widget
This commit adds the `_removeExpense` method to the `_ExpensesState` class in the `Expenses` widget. The method removes the specified `Expense` object from the `_registeredExpenses` list by calling the `setState` method. This enhancement allows users to remove expenses from the list, improving the functionality and user experience of the Expenses widget.